add walks training to BERT

+++ code/jrdf2vec_walks_for_bert.py
@@ -1,11 +1,13 @@
 from utils import run_str
-
+from pathlib import Path
 
 def generate_walks(tmpdir, dataset_file, outname, nproc, depth, no_walks_per_entity, generation_mode="RANDOM_WALKS_DUPLICATE_FREE"):
-    run_str(f"rm -rvf {tmpdir}")
-    run_str(
-        f"java -jar jrdf2vec-1.3-SNAPSHOT.jar -walkDirectory {tmpdir} -graph {dataset_file} -onlyWalks -threads {nproc} -depth {depth} -numberOfWalks {no_walks_per_entity} -walkGenerationMode {generation_mode}")
-    run_str(f"java -jar jrdf2vec-1.3-SNAPSHOT.jar -walkDirectory {tmpdir} -mergeWalks -o {outname}_{generation_mode}_d={depth}_w={no_walks_per_entity}.txt")
+    if not Path(f'{outname}_{generation_mode}_d={depth}_w={no_walks_per_entity}.txt').is_file():
+        run_str(f"rm -rvf {tmpdir}")
+        run_str(
+            f"java -jar jrdf2vec-1.3-SNAPSHOT.jar -walkDirectory {tmpdir} -graph {dataset_file} -onlyWalks -threads {nproc} -depth {depth} -numberOfWalks {no_walks_per_entity} -walkGenerationMode {generation_mode}")
+        run_str(f"java -jar jrdf2vec-1.3-SNAPSHOT.jar -walkDirectory {tmpdir} -mergeWalks -o {outname}_{generation_mode}_d={depth}_w={no_walks_per_entity}.txt")
+    return f"{outname}_{generation_mode}_d={depth}_w={no_walks_per_entity}.txt"
 
 if __name__ == "__main__":
